Address review comment: Move treq_with_authentication someplace better.
