+++ flocker/ca/__init__.py
@@ -10,7 +10,7 @@ __all__ = [
     "KeyAlreadyExistsError", "EXPIRY_20_YEARS",
     "AUTHORITY_CERTIFICATE_FILENAME", "AUTHORITY_KEY_FILENAME",
     "amp_server_context_factory", "rest_api_context_factory",
-    "ControlServicePolicy",
+    "ControlServicePolicy", "treq_with_authentication",
 ]
 
 from ._ca import (
@@ -22,4 +22,5 @@ from ._ca import (
 
 from ._validation import (
     amp_server_context_factory, rest_api_context_factory, ControlServicePolicy,
+    treq_with_authentication,
 )
